refactor: refactoring directory structure
