+++ src/cli/syncManual.ts
@@ -5,10 +5,10 @@
 // Description: CLI script to trigger a manual synchronization job
 
 import { logger } from '../lib/logger.js';
-import { db } from '../db/index.js';
-import { syncOrchestrator } from '../orchestrator/syncOrchestrator.js';
-import { ISyncJobResult } from '../orchestrator/syncJobResult.js';
-import { JobType } from '../enums/db.enums.js';
+import { db } from '../domain/db/impl/sqlite3.js';
+import { syncOrchestrator } from '../domain/orchestrator/impl/syncOrchestrator.js';
+import { ISyncJobResult } from '../domain/orchestrator/syncJobResult.js';
+import { JobType } from '../domain/db/enum/db.enums.js';
 import { asError } from '../lib/utils.js';
 import { StopWatch } from '../lib/stopWatch.js';
